# docs/api-spec.json
{
  "components": {
    "schemas": {
...
      "BuiltinPreset": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "weights": {
            "$ref": "#/components/schemas/WeightSet"
          }
        },
        "required": [
          "name",
          "label",
          "description",
          "weights"
        ],
        "title": "BuiltinPreset",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PresetListResponse": {
        "properties": {
          "builtin": {
            "items": {
              "$ref": "#/components/schemas/BuiltinPreset"
            },
            "title": "Builtin",
            "type": "array"
          },
          "user": {
            "items": {
              "$ref": "#/components/schemas/UserPreset"
            },
            "title": "User",
            "type": "array"
          }
        },
        "required": [
          "builtin",
          "user"
        ],
        "title": "PresetListResponse",
        "type": "object"
      },
      "PresetUpsertRequest": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "weights": {
            "$ref": "#/components/schemas/WeightSet"
          }
        },
        "required": [
          "name",
          "weights"
        ],
        "title": "PresetUpsertRequest",
        "type": "object"
      },
...
      "UserPreset": {
        "properties": {
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "weights": {
            "$ref": "#/components/schemas/WeightSet"
          }
        },
        "required": [
          "name",
          "description",
          "weights"
        ],
        "title": "UserPreset",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WeightSet": {
        "properties": {
          "long": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Long",
            "type": "object"
          },
          "mid": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Mid",
            "type": "object"
          },
          "short": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Short",
            "type": "object"
          }
        },
        "required": [
          "short",
          "mid",
          "long"
        ],
        "title": "WeightSet",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Next.js 前端的 FastAPI 後端，包裝既有 app.* 模組。",
    "title": "台股研究儀表板 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/weight-tuner/presets": {
      "get": {
        "operationId": "list_presets_api_weight_tuner_presets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Presets",
        "tags": [
          "weight-tuner"
        ]
      },
      "post": {
        "operationId": "upsert_preset_api_weight_tuner_presets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreset"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Preset",
        "tags": [
          "weight-tuner"
        ]
      }
    },
...
    "/api/weight-tuner/presets/{name}": {
      "delete": {
        "description": "冪等刪除：刪不存在的 user preset 也回 204。內建 preset 不可刪（delete_preset 丟\nValueError → 422）。",
        "operationId": "delete_preset_api_weight_tuner_presets__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Preset",
        "tags": [
          "weight-tuner"
        ]
      },
      "get": {
        "description": "取單一 user preset 詳情（含 weights）。內建 preset 不在此回傳，請改打 `/presets`.",
        "operationId": "get_preset_api_weight_tuner_presets__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreset"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Preset",
        "tags": [
          "weight-tuner"
        ]
      }
    }
  }
}